move coil package out of download package and into image move CoverViewImageTarget into coil package
