+++ app/src/main/java/eu/kanade/tachiyomi/ui/recents/RecentMangaHolder.kt
@@ -3,7 +3,7 @@ package eu.kanade.tachiyomi.ui.recents
 import android.app.Activity
 import android.view.View
 import eu.kanade.tachiyomi.R
-import eu.kanade.tachiyomi.data.download.coil.loadLibraryManga
+import eu.kanade.tachiyomi.data.image.coil.loadLibraryManga
 import eu.kanade.tachiyomi.data.download.model.Download
 import eu.kanade.tachiyomi.source.LocalSource
 import eu.kanade.tachiyomi.ui.manga.chapter.BaseChapterHolder
