+++ app/src/main/java/eu/kanade/tachiyomi/ui/source/global_search/GlobalSearchMangaHolder.kt
@@ -8,7 +8,7 @@ import coil.request.LoadRequest
 import eu.kanade.tachiyomi.data.database.models.Manga
 import eu.kanade.tachiyomi.ui.base.holder.BaseFlexibleViewHolder
 import eu.kanade.tachiyomi.util.view.visibleIf
-import eu.kanade.tachiyomi.widget.CoverViewTarget
+import eu.kanade.tachiyomi.data.image.coil.CoverViewTarget
 import kotlinx.android.synthetic.main.source_global_search_controller_card_item.*
 
 class GlobalSearchMangaHolder(view: View, adapter: GlobalSearchCardAdapter) :
